Fixes #5 : capinfos crash
For the time being, use a loop with subprocess to interact with capinfos
in case there is a crash for one of the files. This way, we avoid the
wireshark bug. When it is fixed in a majority of installations,
then we can go back to using
`capinfos -MaceS <space-delimited filenames>`